refactor(KFLUXUI-1473): rename formik to formik-base and fix review findings
Rename formik/ to formik-base/ to clarify the distinction from
formik-fields/. Fix icon imports (dist/js to dist/esm), remove wrong
FormEvent cast in BaseToggleField, and fix relative import in
SecretSection.
