+++ src/components/ImportForm/SecretSection/SecretSection.tsx
@@ -17,10 +17,10 @@ import { useFormikContext } from 'formik';
 import { Base64 } from 'js-base64';
 import { IMPORT_SECRET_HELP_TEXT } from '~/consts/secrets';
 import { IfFeature } from '~/feature-flags/hooks';
+import { InputField } from '~/shared/components/formik-base';
+import TextColumnField from '~/shared/components/formik-fields/text-column-field/TextColumnField';
 import { useSecrets } from '../../../hooks/useSecrets';
 import { SecretModel } from '../../../models';
-import { InputField } from '../../../shared/components/formik';
-import TextColumnField from '../../../shared/components/formik-fields/text-column-field/TextColumnField';
 import { useNamespace } from '../../../shared/providers/Namespace';
 import {
   BuildTimeSecret,
