+++ src/components/IntegrationTests/IntegrationTestForm/IntegrationTestSection.tsx
@@ -4,7 +4,7 @@ import { useField } from 'formik';
 import HelpPopover from '~/components/HelpPopover';
 import { LEARN_MORE_ABOUT_INTEGRATION_TESTS } from '~/consts/documentation';
 import { ExternalLink } from '~/shared';
-import { InputField, RadioGroupField, CheckboxField } from '~/shared/components/formik';
+import { InputField, RadioGroupField, CheckboxField } from '~/shared/components/formik-base';
 import { ResourceKind } from '~/types/coreBuildService';
 import { RESOURCE_NAME_REGEX_MSG } from '../../../utils/validation-utils';
 import ContextsField from '../ContextsField';
