+++ src/components/ReleaseService/ReleasePlan/TriggerRelease/TriggerReleaseForm.tsx
@@ -12,7 +12,7 @@ import { PlusCircleIcon } from '@patternfly/react-icons/dist/esm/icons/plus-circ
 import { FormikProps, useField } from 'formik';
 import isEmpty from 'lodash-es/isEmpty';
 import { FilterContextProvider } from '~/components/Filter/generic/FilterContext';
-import { InputField, TextAreaField } from '~/shared/components/formik';
+import { InputField, TextAreaField } from '~/shared/components/formik-base';
 import { getErrorState } from '~/shared/utils/error-utils';
 import { useReleasePlans } from '../../../../../src/hooks/useReleasePlans';
 import PageLayout from '../../../../components/PageLayout/PageLayout';
