+++ src/components/Secrets/SecretsForm/SecretLinkOption.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { SECRET_LINK_OPTION_HELP_TEXT } from '~/consts/secrets';
-import { RadioGroupField } from '~/shared/components/formik';
+import { RadioGroupField } from '~/shared/components/formik-base';
 import { CurrentComponentRef } from '~/types';
 import { SecretForComponentOption } from '~/utils/secrets/secret-utils';
 import HelpPopover from '../../HelpPopover';
