+++ src/components/modal/DeleteResourceModal.tsx
@@ -15,7 +15,7 @@ import {
 } from '@patternfly/react-core';
 import { ModalVariant } from '@patternfly/react-core/deprecated';
 import { Formik } from 'formik';
-import { InputField } from '~/shared/components/formik';
+import { InputField } from '~/shared/components/formik-base';
 import { useNamespace } from '~/shared/providers/Namespace';
 import { K8sQueryDeleteResource } from '../../k8s';
 import { K8sModelCommon, K8sResourceCommon } from '../../types/k8s';
