+++ src/shared/components/formik-base/BaseFields/BaseInputField.tsx
@@ -7,7 +7,7 @@ import {
   TextInputTypes,
   ValidatedOptions,
 } from '@patternfly/react-core';
-import { ExclamationCircleIcon } from '@patternfly/react-icons/dist/js/icons/exclamation-circle-icon';
+import { ExclamationCircleIcon } from '@patternfly/react-icons/dist/esm/icons/exclamation-circle-icon';
 import { useField } from 'formik';
 import { FieldProps } from '../types';
 import { getFieldId } from '../utils';
