+++ src/shared/components/formik-fields/text-column-field/TextColumnItemContent.tsx
@@ -9,7 +9,7 @@ import {
   Tooltip,
 } from '@patternfly/react-core';
 import { MinusCircleIcon, GripVerticalIcon } from '@patternfly/react-icons/dist/esm/icons';
-import { InputField } from '~/shared/components/formik';
+import { InputField } from '~/shared/components/formik-base';
 import {
   TextColumnItemProps,
   TextColumnFieldChildParameterProps,
